+++ main.cpp
@@ -7,7 +7,7 @@
 #define WIDTH 256
 #define HEIGHT 256
 
-int Image[HEIGHT * WIDTH * 3] = {0};
+uint8 Image[HEIGHT * WIDTH * 3] = {0};
 const int KernelImp[3 * 3 * 3 * 3] = {0, 0, 0, 0, 4, 0, 0, 0, 0,
                                       0, 0, 0, 0, 0, 0, 0, 0, 0,
                                       0, 0, 0, 0, 0, 0, 0, 0, 0,
@@ -26,7 +26,7 @@ int main()
     flattenPNG(Image);
 
     Convolution C(KernelImp, 3, 3);
-    int* Y = C.apply(Image, WIDTH, HEIGHT);
+    uint8* Y = C.apply(Image, WIDTH, HEIGHT);
 
     unflattenPNG(Y);
     writePNG("lena2.png");
