+++ png_utils.h
@@ -6,8 +6,8 @@
 #include <ac_fixed.h>
 
 void readPNG (const char *filename);
-void flattenPNG (int *);
-void unflattenPNG (int *I);
+void flattenPNG (uint8 *);
+void unflattenPNG (uint8 *I);
 void writePNG (const char* filename);
 
 #endif // PNG_UTILS_H
